[IMP] Add methods to Deck() class (shuffle)
The new availables methods for Deck() are :

- get_list() : return a Python list of all the Cards in the Deck.
- cut_all_next() : remove all 'next' relation between Cards that compose the Deck.
- shuffle() : shuffle the Deck thanks to random module.

In Card() class, I add a new method to :

- cut_next() : remove the 'next' relationi of a Card.

Why cut_next() and cut_all_next() methods ?
>>> Because when I tried to implement the shuffle method for Deck, I faced an issue.
I taken the list of the cards thanks to get_list() and then I shuffled it, but then, when I tried to remake the Deck the code loop endlessly.
This is because : the Card keep their 'next' relation, even if we shuffle.

So the good way to shuffle the Deck is :
- get the list of the cards
- clear all the 'next' relation
- shuffle the list
- remake the Deck