[IMP] Beggining of Weapon() class
A Weapon is a standard Card() instance with an history of already defended Cards.
To remind, a Weapon can only defend of Card with lower power that the last Card already defended.

The already defended Cards by a Weapon can be see as a linked list, as Deck().
So I make the Weapon class herited of Card class and add a new propertie : history (a linked list --> a Deck() instance).

Methods:
- is_valid_weapon() : in standard rules, only Diamonds Cards could be used as Weapon. I decided to create a class propertie weapon_suits (list) to maybe change that rules and maybe authorize others suits to be weapons (why not after all?).

- get_last_defended_card() : return the last defended Card of the current Weapon (I think this method have a problem when the history is emply (the issues seems comes from Card methods).

- can_defend_on(card) : method that return True if the current Weapon can defend on card, False otherwise.