[IMP] Interaction with Card in Player's class
During the game, the Player have to interact with Cards from the Board.
So I think it's a good idea to implement interaction logic in Player's class.

I make a global interaction method which take a Card in entry.
This method check what kind of Card is this and redirect to the correct interaction method according to that.

The correct interaction methods could be:
- interact_with_monster
- interact_with_potion
- interact_with_weapon

Notes:
I delete my old method 'equip_weapon' to be more efficient with my new logic (-> now it's interact_with_weapon).