[IMP] interact_with_monster() (Player) is more complete
Now, interact_with_monster ask the Player for defend if he has a weapon equiped.
If the Player wants to defend, the attacking Card is add to the Weapon's history.

Notes :
Next, I will have to fix the is_lower() method (Card) because it raises Error when one of the two Cards is None..